// src/app/shared/models/signos-vitales.model.ts
export class SignosVitales {
    
        idMezclaAplicDiaDosis: number;
        numPacTenArtSistolica: number;
        numPacTenArtDiastolica: number;
        numPacFrecCardiaca: number;
        numPacFrecRespiratoria: number;
        numPacTemperatura: number;
        idPaciente: number;
        idUsuarioResponsable: number;
        refPacObs: string;
        cveUsuarioAlta: number;
      

}



export class SignosVitalesPermitidos {
    grupoEdad:String;
        
        numPacTenArtSistolicaMax: number;
        numPacTenArtSistolicaMin: number;
        numPacTenArtDiastolicaMax: number;
        numPacTenArtDiastolicaMin: number;
        numPacFrecCardiacaMax: number;
        numPacFrecCardiacaMin: number;
        numPacFrecRespiratoriaMax: number;
        numPacFrecRespiratoriaMin: number;
        numPacTemperaturaMax: number;
        numPacTemperaturaMin: number;

        numPacTenArtSistolicaMaxLong: number;
        numPacTenArtSistolicaMinLong: number;
        numPacTenArtDiastolicaMaxLong: number;
        numPacTenArtDiastolicaMinLong: number;
        numPacFrecCardiacaMaxLong: number;
        numPacFrecCardiacaMinLong: number;
        numPacFrecRespiratoriaMaxLong: number;
        numPacFrecRespiratoriaMinLong: number;
        numPacTemperaturaMaxLong: number;
        numPacTemperaturaMinLong: number;
       
      

}